Major refactoring, completed Character Class UI
Replaced the entire event listening system with a much more dynamic and fluid alternative: instead of registering Listener classes, you simply pass it a symbol and a block. This allows event listeners to be added to any component from any place without need for class definitions.

Added a FlipPanel, which allows a player to "flip" through child components as if turning pages in a book.

Also completed the Character Class interface, leaving only the Options screen for completion and some basic bug hunting before the game world itself can be started.