Major improvements and documentation for OpenGl::Frustum class
#point_visible? now checks the specified point directly instead of calling #sphere_visible? with radius 0
#sphere_visible? now returns :partial for partial matches.
#cube_visible? now calculates bounding boxes based on object transforms, or uses supplied bounding box (an array of vertices).
#cube_visible? now returns :partial for partial matches.
Added #bounding_box to calculate transformed bounding box from width, height, depth, position, view, up and right vectors. If an object has a camera instance, this is trivial.
Added #poly_visible? to test whether a polygon with an arbitrary number of vertices is in the frustum.
Alias #polygon_visible? for #poly_visible?