Some optimization for MVC components
Most of the slowness was due to mipmapping. I think most likely because the test case I was using involved an oddly shaped image (800x64), which gluBuild2DMipMaps was probably converting into powers of 2. Or maybe it was just too big an image. Anyway, switched it from mipmaps (which were unnecessary) to glTexImage2D and problem solved.